+++ packages/hippo_core/hippo_core/lib/hippo_core.dart
@@ -11,6 +11,7 @@ export 'src/bloc.dart';
 export 'src/data_subject/data_subject.dart';
 export 'src/filter/filter.dart';
 export 'src/filter/filter_controller.dart';
+export 'src/object_store.dart';
 export 'src/store/key_value_store.dart';
 export 'src/store/store_controller.dart';
 export 'src/typedefs/typedefs.dart';